% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gt_data.R
\name{calculateAlleleFreqs_int}
\alias{calculateAlleleFreqs_int}
\title{calculateAlleleFreqs}
\usage{
calculateAlleleFreqs_int(object, missing_data = "none", ...)
}
\arguments{
\item{object}{A S4 object of class vcfR.}

\item{missing_data}{Method to deal with missing data. Options are "remove", "impute", "none". Default is "none". In case of "remove", function needs the additional parameter threshold, which is the fraction of missing data in a variant taht is still acceptable. In case of "impute" the function needs the additional parameters "method", with which the imputation method can chosen. See \code{\link{imputeMissingData}}}
}
\value{
A S4 object of the same class but with following slots added:
\itemize{
\item allele_freqs (data frame)
}
}
\description{
Calculate allele frequencies from a vcfR object. Will also add the slots ploidy and sep_gt, if not already present.
}
\examples{
data("real", package = "GenoPop")
vcf <- calculateAlleleFreqs_int(real, missing_data = "none")
vcf <- calculateAlleleFreqs_int(real, missing_data = "remove", threshold = 0.1)
vcf <- calculateAlleleFreqs_int(real, missing_data = "impute", method = "mean")
head(vcf@allele_freqs)

}
