% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gt_data.R
\name{calculatePloidyAndSepGT}
\alias{calculatePloidyAndSepGT}
\title{calculatePloidyAndSepGT}
\usage{
calculatePloidyAndSepGT(object)
}
\arguments{
\item{object}{A S4 object of class vcfR.}
}
\value{
A S4 object of the same class but with following slots added:
\itemize{
\item ploidy (integer)
\item sep_gt (matrix)
}
}
\description{
Calculate ploidy levels and separate the genotypes into a matrix according to the ploidy. This function is mainly used internally. If you want to these oprations, its easier to start here :  \code{\link{calculateAlleleFreqs}}
}
\examples{
data("real", package = "GenoPop")
vcf <- calculatePloidyAndSepGT(real)
vcf@ploidy
head(vcf@sep_gt)

}
