% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gt_data.R
\name{calculatePloidyAndSepGT_int}
\alias{calculatePloidyAndSepGT_int}
\title{calculatePloidyAndSepGT}
\usage{
calculatePloidyAndSepGT_int(object)
}
\arguments{
\item{object}{A S4 object of class vcfR.}
}
\value{
A S4 object of the same class but with following slots added:
\itemize{
\item ploidy (integer)
\item sep_gt (matrix)
}
}
\description{
Calculate ploidy levels and separate the genotypes into a matrix according to the ploidy.
}
\examples{
data("real", package = "GenoPop")
vcf <- calculatePloidyAndSepGT_int(real)
vcf@ploidy
head(vcf@sep_gt)

}
