% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dav}
\alias{dav}
\title{The test data set called 'dav'}
\format{
A myVcfR object
}
\usage{
data("dav", package = "GenoPop")
}
\description{
This is the real data set, but seperated by population (mys & dav) with missing data imputed (rf) and allele frequencies calculated.
}
\examples{
\dontrun{
data("dav", package = "GenoPop")
head(dav@imp_gt)
}
}
\keyword{datasets}
