% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{Dxy}
\alias{Dxy}
\title{Calculate Average Nucleotide Differences (Dxy) from VCF File in Batch or Window Mode}
\usage{
Dxy(
  vcf_path,
  pop1_individuals,
  pop2_individuals,
  seq_length,
  batch_size = 10000,
  threads = 1,
  write_log = FALSE,
  logfile = "log.txt",
  window_size = NULL,
  skip_size = NULL,
  exclude_ind = NULL
)
}
\arguments{
\item{vcf_path}{Path to the VCF file.}

\item{pop1_individuals}{Vector of individual names belonging to the first population.}

\item{pop2_individuals}{Vector of individual names belonging to the second population.}

\item{seq_length}{Length of the sequence in number of bases, including monomorphic sites (used in batch mode only).}

\item{batch_size}{The number of variants to be processed in each batch
(used in batch mode only, default of 10,000 should be suitable for most use cases).}

\item{threads}{Number of threads to use for parallel processing.}

\item{write_log}{Logical, indicating whether to write progress logs.}

\item{logfile}{Path to the log file where progress will be logged.}

\item{window_size}{Size of the window for windowed analysis in base pairs (optional).
When specified, \code{skip_size} must also be provided.}

\item{skip_size}{Number of base pairs to skip between windows (optional).
Used in conjunction with \code{window_size} for windowed analysis.}

\item{exclude_ind}{Optional vector of individual IDs to exclude from the analysis.
If provided, the function will remove these individuals from the genotype matrix
before applying the custom function. Default is NULL, meaning no individuals are excluded.}
}
\value{
In batch mode (no window_size or skip_size provided): The average number of nucleotide substitutions per site between the individuals of two populations (Dxy).
In window mode (window_size and skip_size provided): A data frame with columns 'Chromosome', 'Start', 'End', and 'Dxy', representing the average nucleotide differences within each window.
}
\description{
This function calculates the average number of nucleotide differences per site (Dxy) between two populations from a VCF file, processing the file either in batches or specified windows across the genome for efficient memory usage.
}
\examples{
vcf_path <- "path/to/your/vcf/file"
pop1_individuals <- c("8449", "8128", "8779")
pop2_individuals <- c("8816", "8823", "8157")
total_sequence_length <- 265392  # Total length of the sequence
# Batch mode example
dxy_value <- Dxy(vcf_path, pop1_individuals, pop2_individuals, total_sequence_length)
# Window mode example
dxy_windows <- Dxy(vcf_path, pop1_individuals, pop2_individuals, seq_length = total_sequence_length, window_size = 100000, skip_size = 50000)

}
