% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gt_data.R
\name{filterBiallelicSNPs_int}
\alias{filterBiallelicSNPs_int}
\title{filterBiallelicSNPs}
\usage{
filterBiallelicSNPs_int(object)
}
\arguments{
\item{object}{A S4 object of class vcfR.}
}
\value{
A S4 object of the same class, but complex and multiallelic variantes are removed from the @fix and @gt slots.
}
\description{
Filter for only biallelic SNPs in the data set.
}
\examples{
data("real", package = "GenoPop")
vcf <- filterBiallelicSNPs_int(real)

}
