% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{FixedSites}
\alias{FixedSites}
\title{FixedSites}
\usage{
FixedSites(object)
}
\arguments{
\item{object}{An S4 object of type GPvcfR. Allele frequencies must be present.}
}
\value{
The number of fixed sites.
}
\description{
Count the number of sites fixed for the alternative allele ("1").
}
\examples{
data("real", package = "GenoPop")
vcf <- calculateAlleleFreqs(real, missing_data = "impute", method = "mean")
FixedSites(vcf)

}
