% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{Fst}
\alias{Fst}
\title{Calculate Fst from VCF File}
\usage{
Fst(
  vcf_path,
  pop1_individuals,
  pop2_individuals,
  weighted = FALSE,
  batch_size = 10000,
  threads = 1,
  write_log = FALSE,
  logfile = "log.txt"
)
}
\arguments{
\item{vcf_path}{Path to the VCF file.}

\item{pop1_individuals}{Vector of individual names belonging to the first population.}

\item{pop2_individuals}{Vector of individual names belonging to the second population.}

\item{weighted}{Logical, whether weighted Fst or mean Fst is returned (Default = FALSE (mean Fst is returned)).}

\item{batch_size}{The number of variants to be processed in each batch
(default of 10,000 should be suitable for most use cases).}

\item{threads}{Number of threads to use for parallel processing.}

\item{write_log}{Logical, indicating whether to write progress logs.}

\item{logfile}{Path to the log file where progress will be logged.}
}
\value{
Fst value (either mean or weighted).
}
\description{
This function calculates the fixation index (Fst) between two populations from a VCF file using the method of Weir and Cockerham (1984). It processes the file in batches for efficient memory usage.
The result can be either the mean or weighted Fst based on the number of non-missing chromosomes.
}
\examples{
vcf_path <- "path/to/your/vcf/file"
pop1_individuals <- c("8449", "8128", "8779")
pop2_individuals <- c("8816", "8823", "8157")
fst_value <- Fst(vcf_path, pop1_individuals, pop2_individuals, weighted = TRUE)

}
