% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{GenoPop}
\alias{GenoPop}
\alias{_PACKAGE}
\title{GenoPop}
\description{
A R package to perform several population genomics analyses directly on whole genome vcf files.
}
\details{
Most important features are:
\itemize{
\item Reading genotype data ready to use from vcf files using the vcfR package.
\item Different methods to impute or remove missing data from the genotype matrix.
\item Calculation of several commonly used population genomics metrics from the genotype data.
\item Window based analyses with those different metrics.
\item Parallelization and optimization of heavy tasks to enable processing of whole genomes (imputation and window based analyses).
\item Writing processed data back to file in vcf format.
}
\subsection{Installation Instructions for GenoPop}{
\subsection{Prerequisites}{

Before installing GenoPop, make sure you have R installed on your system. You can download and install R from \href{https://cran.r-project.org/}{CRAN}.
}

\subsection{Installing GenoPop from GitHub}{

To install the GenoPop package directly from GitHub, you will need the \code{devtools} package in R. If you don't have \code{devtools} installed, you can install it by running the following command in R:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{install.packages("devtools")
}\if{html}{\out{</div>}}

Once devtools is installed, you can install GenoPop using the install_github function. Run the following commands in your R console:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{library(devtools)
install_github("MiGurke/GenoPop")
}\if{html}{\out{</div>}}

Then load the package..

\if{html}{\out{<div class="sourceCode R">}}\preformatted{library(GenoPop)
}\if{html}{\out{</div>}}

and your ready to go!
}

\subsection{Dependencies}{

For proper compression of vcf's newly generated, you need to have tabix installed on your machine. All other dependencies are supposed the handled by R itself.
}

}

\subsection{Getting started}{

To get your vcf formated data into R use the reading function of the vcfR package:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{vcf <- read.vcfR( "example.vcf", verbose = FALSE )
}\if{html}{\out{</div>}}

It can read compressed and umcompressed vcf files.

From there a good point to start your analysis is to use the calculateAlleleFreqs function to just populate the most important data and information slots from you vcf file automatically.

\if{html}{\out{<div class="sourceCode R">}}\preformatted{example <- calculateAlleleFreqs(vcf)
}\if{html}{\out{</div>}}

This will give you a nicely formatted genotype matrix (example@sep_gt), calculated allele frequencies (example@allele_freqs) and some interesting stats about the amount of missing data in you data set (example@missing_data).

From there it depends on you how to continue. You can fix some issue with missing data by imputing or removing it, directly calculate some stats like Fst, Pi, and the site frequency spectrum, or even do a window based analysis. Just have a look at the available functions in the man pages.

If you need further assistance or got suggestions for this package, feel free to open an issue on the GenoPop GitHub repo or contact me in any other way.
}
}
\author{
\strong{Maintainer}: Marie (Mick) Gurke \email{margurke@gmail.com} (\href{https://orcid.org/0000-0001-9901-424X}{ORCID})

}
