% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{Heterozygosity}
\alias{Heterozygosity}
\title{Heterozygosity}
\usage{
Heterozygosity(object)
}
\arguments{
\item{object}{An S4 object of type myVcfR. Allele frequencies must be present.}
}
\value{
A list containing observed and expected heterozygosity.
}
\description{
This function calculates the observed and expected heterozygosity of a population.
}
\examples{
data("mys", package = "GenoPop")
Heterozygosity(mys)

}
