% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{imputeMissingData}
\alias{imputeMissingData}
\title{imputeMissingData}
\usage{
imputeMissingData(object, method = "mean", ...)
}
\arguments{
\item{object}{A S4 object of class vcfR.}

\item{method}{Method used for missing data imputation. Available are "kNN", "rf", and "mean". (Default = "mean").}

\item{...}{Additional parameters for different imputation methods. For more info look up the documentation of them:\code{\link{meanImputation}}, \code{\link{kNNImputation}}, \code{\link{rfImputation}}.}
}
\value{
A S4 object of the same class, but the slot imp_gt is now filled with the imputed genotype matrix.
}
\description{
Impute missing variants in genotype data stored in vcfR object.
}
\examples{
data("real", package = "GenoPop")
vcf <- imputeMissingData(real, method = "mean")

}
