% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{knn_imputeR}
\alias{knn_imputeR}
\title{knn_imputeR}
\usage{
knn_imputeR(data, k)
}
\arguments{
\item{data}{NumericMatrix, the data matrix.}

\item{k}{Integer, the number of neighbors.}
}
\value{
NumericMatrix, the imputed data matrix.
}
\description{
Description...
}
\examples{
knn_imputeR(matrix, k = 3)
}
