% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{meanImputation}
\alias{meanImputation}
\title{meanImputation}
\usage{
meanImputation(
  vcf_path,
  output_vcf,
  batch_size = 10000,
  threads = 1,
  write_log = FALSE,
  logfile = "log.txt",
  mode = "variant"
)
}
\arguments{
\item{mode}{Means are calculated either by "variant" or by "individual". (Default: "variant")}

\item{sep_gt}{separated genotype matrix from myvcfR object.}
}
\value{
Matrix with imputed missing data.
}
\description{
Use the mean over each variant or individual to replace missing data with that mean. Mean is rounded to keep genotype integers, so this just corresponds to the most often occuring genotyp. If you want to use this algorithm on your data, please use the \code{\link{imputeMissingData}} function which will do the operation on GPvcfR object.
}
\examples{
example_matrix <- matrix(c("0", "1", ".", "1", "1", ".", "0", "0", ".", "1", "1", ".", "0", "0", "0", "0", ".", "1", "0", "0", ".", "1", "1", ".", "0"), nrow = 5, byrow = TRUE)
meanImputation(example_matrix, mode ="variant")
meanImputation(example_matrix, mode ="individual")

}
