% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mys}
\alias{mys}
\title{The test data set called 'mys'}
\format{
A GPvcfR object
}
\usage{
data("mys", package = "GenoPop")
}
\description{
This is the real data set, but separated by population (mys & dav) with missing data imputed (rf) and allele frequencies calculated.
}
\examples{
\dontrun{
data("mys", package = "GenoPop")
head(mys@imp_gt)
}
}
\keyword{datasets}
