% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{ObservedHeterozygosity}
\alias{ObservedHeterozygosity}
\title{Calculate Observed Heterozygosity from VCF File}
\usage{
ObservedHeterozygosity(
  vcf_path,
  batch_size = 10000,
  threads = 1,
  write_log = FALSE,
  logfile = "log.txt"
)
}
\arguments{
\item{vcf_path}{Path to the VCF file.}

\item{batch_size}{The number of variants to be processed in each batch
(default of 10,000 should be suitable for most use cases).}

\item{threads}{Number of threads to use for parallel processing.}

\item{write_log}{Logical, indicating whether to write progress logs.}

\item{logfile}{Path to the log file where progress will be logged.}
}
\value{
Observed heterozygosity averaged over all loci.
}
\description{
This function calculates the observed heterozygosity (Ho) for each variant in a VCF file.
It processes the file in batches for efficient memory usage.
}
\examples{
vcf_path <- "path/to/your/vcf/file"
Ho <- ObservedHeterozygosity(vcf_path)

}
