% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{PolymorphicSites}
\alias{PolymorphicSites}
\title{PolymorphicSites}
\usage{
PolymorphicSites(object)
}
\arguments{
\item{object}{An S4 object of type GPvcfR. Allele frequencies must be present.}
}
\value{
The number of polymorphic sites.
}
\description{
Count the number of polymorphic sites in the data set (aka. sites not fixed for the alternative allele).
}
\examples{
data("real", package = "GenoPop")
vcf <- calculateAlleleFreqs(real, missing_data = "impute", method = "mean")
PolymorphicSites(vcf)

}
