% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{PrivateAlleles}
\alias{PrivateAlleles}
\title{PrivateAlleles}
\usage{
PrivateAlleles(objects)
}
\arguments{
\item{objects}{A list of S4 objects of type myVcfR, each representing a different population. Allele frequencies must be present.}
}
\value{
A list containing the number of private alleles for each population.
}
\description{
Count the number of private alleles in each of several populations. These are alleles found exclusively in one population and not in the others.
}
\examples{
data("mys", package = "GenoPop")
data("dav", package = "GenoPop")
PrivateAlleles(list(mys, dav))

}
