% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{real}
\alias{real}
\title{The test data set called 'real'}
\format{
A vcfR object.
}
\usage{
data("real", package = "GenoPop")
}
\description{
These are first couple of thousand lines of a vcf file from real world data. It contains genotype information of 16 individuals of two bat species.
}
\examples{
\dontrun{
data("real", package = "GenoPop")
head(vcf@gt)
}
}
\keyword{datasets}
