% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{rmMissingData_int}
\alias{rmMissingData_int}
\title{rmMissingData}
\usage{
rmMissingData_int(object, threshold = 0.1)
}
\arguments{
\item{object}{A S4 object of class vcfR.}

\item{threshold}{Fraction of missing individuals per variant that is still accepted. Default: 0.1}
}
\value{
A S4 object of the same class, but without variants that did not meet the missingness threshold. In addition, the slot "missing_data" will be added to the object. It contains a list with information about the removed variants and the missingness per variant and individual.
}
\description{
Remove variants from vcfR object with to much missing data.
}
\examples{
data("real", package = "GenoPop")
vcf <- rmMissingData_int(real, 0.1)

}
