% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{SegregatingSites}
\alias{SegregatingSites}
\title{SegregatingSites}
\usage{
SegregatingSites(object)
}
\arguments{
\item{object}{An S4 object of type myVcfR. Allele frequencies must be present.}
}
\value{
The number of segregating sites.
}
\description{
Count the number of segregating sites in the data set.
}
\examples{
data("real", package = "GenoPop")
vcf <- calculateAlleleFreqs(real, missing_data = "impute", method = "mean")
SegregatingSites(vcf)

}
