% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\title{The test data set called 'sim'}
\format{
VCF File format
}
\usage{
data("sim", package = "GenoPop")
}
\description{
These are first couple of thousand lines of a vcf file from a simulated data set that was created by msprime. It also contains genotype information of 16 individuals.
}
\examples{
\dontrun{
data("sim", package = "GenoPop")
head(vcf@gt)
}
}
\keyword{datasets}
