% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{SingeltonSites}
\alias{SingeltonSites}
\title{SingeltonSites}
\usage{
SingeltonSites(object)
}
\arguments{
\item{object}{An S4 object of type GPvcfR. Allele frequencies must be present.}
}
\value{
The number of singelton sites.
}
\description{
Count the number of singelton sites in the data set. These are sites where a minor allele occurs only once in the sample.
}
\examples{
data("real", package = "GenoPop")
vcf <- calculateAlleleFreqs(real, missing_data = "impute", method = "mean")
SingeltonSites(vcf)

}
