% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{TajimasD}
\alias{TajimasD}
\title{Calculate Tajima's D from VCF File}
\usage{
TajimasD(
  vcf_path,
  seq_length,
  batch_size = 10000,
  threads = 1,
  write_log = FALSE,
  logfile = "log.txt"
)
}
\arguments{
\item{vcf_path}{Path to the VCF file.}

\item{seq_length}{Total length of the sequence in number of bases.}

\item{batch_size}{The number of variants to be processed in each batch
(default of 10,000 should be suitable for most use cases).}

\item{threads}{Number of threads to use for parallel processing.}

\item{write_log}{Logical, indicating whether to write progress logs.}

\item{logfile}{Path to the log file where progress will be logged.}
}
\value{
Tajima's D value.
}
\description{
This function calculates Tajima's D statistic for a given dataset. It processes the file in batches for efficient memory usage.
Tajima's D is a measure for neutrality based on the difference between the number of segregating sites and the average number of nucleotide differences.
}
\examples{
vcf_path <- "path/to/your/vcf/file"
total_sequence_length <- 265392  # Total length of the sequence
tajimas_d <- TajimasD(vcf_path, total_sequence_length)

}
