% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{WattersonsTheta}
\alias{WattersonsTheta}
\title{Calculate Watterson's Theta from VCF File}
\usage{
WattersonsTheta(
  vcf_path,
  seq_length,
  batch_size = 10000,
  threads = 1,
  write_log = FALSE,
  logfile = "log.txt"
)
}
\arguments{
\item{vcf_path}{Path to the VCF file.}

\item{seq_length}{The length of the sequence in the data set.}

\item{batch_size}{The number of variants to be processed in each batch
(default of 10,000 should be suitable for most use cases).}

\item{threads}{Number of threads to use for parallel processing.}

\item{write_log}{Logical, indicating whether to write progress logs.}

\item{logfile}{Path to the log file where progress will be logged.}
}
\value{
Watterson's theta value normalized by the sequence length.
}
\description{
This function calculates Watterson's Theta, a measure for neutrality, from a VCF file. It processes the file in batches for efficient memory usage.
The metric will be normalized by the sequence length to make it comparable between datasets.
}
\examples{
vcf_path <- "path/to/your/vcf/file"
total_sequence_length <- 265392  # Total length of the sequence
wattersons_theta <- WattersonsTheta(vcf_path, total_sequence_length)

}
