% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gt_data.R
\name{writeVCF}
\alias{writeVCF}
\title{writeVCF}
\usage{
writeVCF(object, file_path, use_imputed = TRUE, bgzip = FALSE)
}
\arguments{
\item{object}{A GPvcfR object containing the VCF data.}

\item{file_path}{Path to the output VCF file.}

\item{use_imputed}{Logical, indicating whether to use the imputed genotypes if available. If FALSE or imputed genotypes are not available, separated genotypes will be used.}

\item{bgzip}{Logical, indicating whether to bgzip the output file (requires tabix to be installed).}
}
\value{
Invisible NULL. The function is called for its side effect of writing a file.
}
\description{
Writes a new VCF file to disk using the imputed or separated genotypes from a GPvcfR object.
}
\examples{
data("mys", package = "GenoPop")
writeVcf(mys, "output.vcf", use_imputed = TRUE)

}
