% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gt_data.R
\name{calculateAlleleFreqs}
\alias{calculateAlleleFreqs}
\title{Calculate Allele Frequencies from Genotype Matrix}
\usage{
calculateAlleleFreqs(sep_gt)
}
\arguments{
\item{sep_gt}{Genotype matrix similar to the \verb{@sep_gt} slot of a \code{vcfR} object.}
}
\value{
A data frame containing allele frequencies for each variant.
}
\description{
This function calculates allele frequencies from a genotype matrix (sep_gt) for each variant.
It is designed to be used within the batch or window processing framework of \code{process_vcf_in_batches} and \code{process_vcf_in_windows}.
}
\keyword{internal}
