% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gt_data.R
\name{calculateWindowedMetric}
\alias{calculateWindowedMetric}
\title{calculateWindowedMetric}
\usage{
calculateWindowedMetric(
  object,
  metricFunction,
  window_size = 1000,
  step_size = 0
)
}
\arguments{
\item{object}{An S4 object of type myVcfR.}

\item{window_size}{The size of the window for which Pi is calculated. (Default = 1000)}

\item{step_size}{The size of the step in between windows. (Default = 0)}
}
\value{
The nucleotide diversity (pi) per window in data frame.
}
\description{
Calculate the average number of nucleotide differences per site between two sequences, for windows of specified length. The algorithm used for this is equivalent to the one used in vcftools --window-pi (https://vcftools.sourceforge.net/man_latest.html).
}
\examples{
data("mys", package = "GenoPop")
WindowedPi(mys)

}
