% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{Dxy}
\alias{Dxy}
\title{Calculate Average Nucleotide Differences (Dxy) from VCF File}
\usage{
Dxy(
  vcf_path,
  pop1_individuals,
  pop2_individuals,
  seq_length,
  batch_size = 10000,
  threads = 1,
  write_log = FALSE,
  logfile = "log.txt"
)
}
\arguments{
\item{vcf_path}{Path to the VCF file.}

\item{pop1_individuals}{Vector of individual names belonging to the first population.}

\item{pop2_individuals}{Vector of individual names belonging to the second population.}

\item{seq_length}{Length of the sequence in number of bases, including monomorphic sites.}

\item{batch_size}{The number of variants to be processed in each batch
(default of 10,000 should be suitable for most use cases).}

\item{threads}{Number of threads to use for parallel processing.}

\item{write_log}{Logical, indicating whether to write progress logs.}

\item{logfile}{Path to the log file where progress will be logged.}
}
\value{
The average number of nucleotide substitutions per site between the individuals of two populations (Dxy).
}
\description{
This function calculates the average number of nucleotide differences per site (Dxy) between two populations from a VCF file.  Nei & Li, 1979 (https://doi.org/10.1073/pnas.76.10.52699).
Handling missing alleles at one site is equivalent to Korunes & Samuk, 2021 ( https://doi.org/10.1111/1755-0998.13326), but for simplicity assuming that completely missing sites are invariant sites, which will underestimate Dxy.
Otherwise this would only function with VCF files that include all monomorphic sites, which may be unpractical given common data sets.
If you happen to know the number of missing sites vs the number of monomorphic sites, please just use the number of monomorphic + the number of polymorphic sites as the sequence length to the the most accurate estimation of Dxy.
}
\examples{
vcf_path <- "path/to/your/vcf/file"
pop1_individuals <- c("8449", "8128", "8779")
pop2_individuals <- c("8816", "8823", "8157")
total_sequence_length <- 265392  # Total length of the sequence
dxy_value <- Dxy(vcf_path, pop1_individuals, pop2_individuals, total_sequence_length)

}
