% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{Dxy}
\alias{Dxy}
\title{Dxy}
\usage{
Dxy(object, pop_assignments, seq_length)
}
\arguments{
\item{object}{A GPvcfR object.}

\item{pop_assignments}{A named vector with elements being the population names and names being the individual names.}

\item{seq_length}{Length of the sequence in number of bases, including monomorphic sites.}
}
\value{
The average number of nucleotide substitutions per sitej between the individuals of two populations (Dxy).
}
\description{
Calculate the average number of nucleotide differences per site (Dxy) between two populations. Nei & Li, 1979 (https://doi.org/10.1073/pnas.76.10.52699).
Handling missing alleles at one site is equivalent to Korunes & Samuk, 2021 ( https://doi.org/10.1111/1755-0998.13326), but for simplicity assuming that completely missing sites are invariant sites, which will underestimate Dxy.
Otherwise this would only function with VCF files that include all monomorphic sites, which may be unpractical given common data sets.
If you happen to know the number of missing sites vs the number of monomorphic sites, please just use the number of monomorphic + the number of polymorphic sites as the sequence length to the the most accurate estimation of Dxy.
}
\examples{
data("mys", package = "GenoPop")
mys1 <- c("8449", "8128", "8779")
mys2 <- c("8816", "8823", "8157")
individuals <- c(mys1, mys2)
pop_names <- c(rep("pop1", length(mys1)), rep("pop2", length(mys2)))
pop_assignments <- setNames(pop_names, individuals)
Dxy(mys, pop_assignments, 265392)

}
