% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{ExpectedHeterozygosity}
\alias{ExpectedHeterozygosity}
\title{Expected Heterozygosity}
\usage{
ExpectedHeterozygosity(object)
}
\arguments{
\item{object}{An S4 object of type myVcfR. Allele frequencies must be present.}
}
\value{
Expected heterozygosity.
}
\description{
This function calculates the expected heterozygosity of a population.
}
\examples{
data("mys", package = "GenoPop")
ExpectedHeterozygosity(mys)

}
