% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{Fst}
\alias{Fst}
\title{Fst}
\usage{
Fst(object, pop_assignments, weighted = FALSE)
}
\arguments{
\item{object}{A GPvcfR object.}

\item{pop_assignments}{A named vector. Elements are the population names and names are the individual name.}

\item{weighted}{Logical, wether weighted Fst or mean Fst is returned. (Default = FALSE (mean Fst is returned))}
}
\value{
Fst value.
}
\description{
Calculate the mean or weighted (by number of non missing chromsomes) fixiation index (Fst) from two populations in a list of GPvcfR objects using the method of Weir and Cockerham (1984).
}
\examples{
mys1 <- c("8449", "8128", "8779")
mys2 <- c("8816", "8823", "8157")

individuals <- c(mys1, mys2)
pop_names <- c(rep("mys1", length(mys1)), rep("mys2", length(mys2)))
pop_assignments <- setNames(pop_names, individuals)

data("mys", package = "GenoPop")
Fst(mys, pop_assignments)

}
