% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{knn_imputeR}
\alias{knn_imputeR}
\title{knn_imputeR}
\usage{
knn_imputeR(data, k)
}
\arguments{
\item{data}{NumericMatrix, the data matrix.}

\item{k}{Integer, the number of neighbors.}
}
\value{
NumericMatrix, the imputed data matrix.
}
\description{
This is just the wrapper function for the kNN imputation algorithm that is written in C++. Please use the function kNNImputation().
}
\examples{
example_matrix <- matrix(c(0, 1, NA, 1, 1, NA, 0, 0, NA, 1, 1, NA, 0, 0, 0, 0, NA, 1, 0, 0, NA, 1, 1, NA, 0), nrow = 5, byrow = TRUE)
knn_imputeR(example_matrix, k = 3)
}
