% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{kNNImputation}
\alias{kNNImputation}
\title{kNNImputation}
\usage{
kNNImputation(sep_gt, k = 3)
}
\arguments{
\item{matrix}{A seperated genotype matrix from a myvcfR object.}
}
\value{
A number, sum of x and y.
}
\description{
kNNImputation
}
\examples{
kNNImputation(matrix)

}
