% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{kNNImputation}
\alias{kNNImputation}
\title{kNNImputation}
\usage{
kNNImputation(
  sep_gt,
  k = 3,
  chunk_size = 1000,
  write_log = FALSE,
  logfile = "logfile.txt"
)
}
\arguments{
\item{sep_gt}{A seperated genotype matrix from a myvcfR object.}

\item{k}{Number of nearest neighbours used for imputation, default: 3.}

\item{chunk_size}{Number of variants analyzed in on batch in the parallelization. Default: 1000. Increasing this might improve accuracy, but will substantially increase running time.}

\item{write_log}{Logical, whether a log file of the process should be written to disk. This is adviced for imputing large data sets.}

\item{logfile}{Name of the log file, if write_log is true.}
}
\value{
A seperated genotype matrix from a myvcfR object, but with imputed missing values.
}
\description{
kNNImputation
}
\examples{
example_matrix <- matrix(c("0", "1", ".", "1", "1", ".", "0", "0", ".", "1", "1", ".", "0", "0", "0", "0", ".", "1", "0", "0", ".", "1", "1", ".", "0"), nrow = 5, byrow = TRUE)
kNNImputation(example_matrix, k = 3, chunk_size = 1000)
}
