% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{meanImputation}
\alias{meanImputation}
\title{meanImputation}
\usage{
meanImputation(sep_gt)
}
\arguments{
\item{matrix}{Seperated genotype matrix from myvcfR object.}
}
\value{
Matrix with imputed missing data.
}
\description{
meanImputation
}
