% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{meanImputation}
\alias{meanImputation}
\title{meanImputation}
\usage{
meanImputation(
  vcf_path,
  output_vcf,
  batch_size = 10000,
  threads = 1,
  write_log = FALSE,
  logfile = "log.txt",
  mode = "variant"
)
}
\arguments{
\item{vcf_path}{Path to the input VCF file.}

\item{output_vcf}{Path for the output VCF file with imputed data.}

\item{batch_size}{Number of variants to process per batch (default: 10000).}

\item{threads}{Number of threads to use for parallel processing (default: 1).}

\item{write_log}{If TRUE, writes a log of the imputation process (default: FALSE).}

\item{logfile}{Path to the log file, used if \code{write_log} is TRUE.}

\item{mode}{Method for calculating means, either "variant" or "individual" (default: "variant").}
}
\value{
Path to the output VCF file with imputed data.
}
\description{
Imputes missing data in VCF files using mean genotypes, either by variant or by individual. This function is designed for large-scale genomic datasets and utilizes efficient batch processing and parallel computing to handle VCF files directly. The function reads a VCF file in batches, performs mean imputation, and writes the imputed data to a new VCF file.
}
\details{
The function operates in two modes: "variant" and "individual". In "variant" mode, it calculates the mean genotype for each variant across all individuals and replaces missing values with this mean. In "individual" mode, it calculates the mean genotype for each individual across all variants and replaces missing values with this mean. The means are rounded to maintain integer genotypes. The imputed data is written to a new VCF file, maintaining the original format.
}
\examples{
meanImputation("/path/to/input.vcf",
               "/path/to/output.vcf",
               batch_size = 1000,
               threads = 5,
               write_log = TRUE,
               logfile = "/path/to/logfile.txt",
               mode = "variant")

}
