% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{ObservedHeterozygosity}
\alias{ObservedHeterozygosity}
\title{Observed Heterozygosity}
\usage{
ObservedHeterozygosity(object)
}
\arguments{
\item{object}{An S4 object of type GPvcfR.}
}
\value{
Observed heterozygosity.
}
\description{
This function calculates the observed heterozygosity in a population.
}
\examples{
data("mys", package = "GenoPop")
ObservedHeterozygosity(mys)

}
