% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{OneDimSFS}
\alias{OneDimSFS}
\title{OneDimSFS}
\usage{
OneDimSFS(object, folded = FALSE)
}
\arguments{
\item{object}{A S4 object of type myVcfR. Allele frequencies and genotype matrix must be present.}

\item{folded}{Logical, deciding if folded (TRUE) or unfolded (FALSE) SFS is returned. For the unfolded it is assumed that the genotype "0" represents the ancestral state in the data. (Default is unfolded (FALSE).)}
}
\value{
Site frequency spectrum as a named vector
}
\description{
Calculate a one dimensional site frequency spectrum from an myVcfR object.
}
\examples{
data("mys", package = "GenoPop")
OneDimSFS(mys, folded = FALSE)

}
