% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{Pi}
\alias{Pi}
\title{Pi}
\usage{
Pi(object, seq_length)
}
\arguments{
\item{object}{An S4 object of type myVcfR.}

\item{seq_length}{Length of the sequence in number of bases. Must be provided to accurately work with all monomorphic sites, including those monomorphic for the reference, which are generally not included in a vcf.}
}
\value{
The nucleotide diversity (pi) per window in data frame.
}
\description{
Calculate the average number of nucleotide differences per site between two sequences. The formula used for this is equivalent to the one used in vcftools --window-pi (https://vcftools.sourceforge.net/man_latest.html).
}
\examples{
data("mys", package = "GenoPop")
Pi(mys, 265392)

}
