% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{PrivateAlleles}
\alias{PrivateAlleles}
\title{PrivateAlleles}
\usage{
PrivateAlleles(object, pop_assignments)
}
\arguments{
\item{object}{A GPvcfR object.}

\item{pop_assignments}{A named vector. Elements are the population names and names are the individual name.}
}
\value{
A list containing the number of private alleles for each population.
}
\description{
Function to calculate the number of private alleles in  two populations.
}
\examples{
mys1 <- c("8449", "8128", "8779")
mys2 <- c("8816", "8823", "8157")

individuals <- c(mys1, mys2)
pop_names <- c(rep("mys1", length(mys1)), rep("mys2", length(mys2)))
pop_assignments <- setNames(pop_names, individuals)

data("mys", package = "GenoPop")
PrivateAlleles(mys, pop_assignments)

}
