% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{rfImputation}
\alias{rfImputation}
\title{rfImputation}
\usage{
rfImputation(
  sep_gt,
  maxiter = 10,
  ntree = 100,
  chunk_size = 1000,
  write_log = FALSE,
  logfile = "logfile.txt"
)
}
\arguments{
\item{sep_gt}{A seperated genotype matrix from a myvcfR object.}

\item{maxiter}{The number of improvement iterations the random forest algorithm (missForest) runs.}

\item{ntree}{The number of decision trees in the random forest.}

\item{chunk_size}{Number of variants analyzed in on batch in the parallelization. Default: 1000. Increasing this might improve accuracy, but will substantially increase running time.}

\item{write_log}{Logical, whether a log file of the process should be written to disk. This is adviced for imputing large data sets.}

\item{logfile}{Name of the log file, if write_log is true.}
}
\value{
A seperated genotype matrix from a myvcfR object, but with imputed missing values.
}
\description{
rfImputation
}
\examples{
example_matrix <- matrix(c("0", "1", ".", "1", "1", ".", "0", "0", ".", "1", "1", ".", "0", "0", "0", "0", ".", "1", "0", "0", ".", "1", "1", ".", "0"), nrow = 5, byrow = TRUE)
rfImputation(example_matrix, maxiter = 10, ntree = 100, chunk_size = 1000)

}
