% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{rfImputation}
\alias{rfImputation}
\title{rfImputation}
\usage{
rfImputation(
  vcf_path,
  output_vcf,
  batch_size = 1000,
  maxiter = 10,
  ntree = 100,
  threads = 1,
  write_log = FALSE,
  logfile = "log.txt"
)
}
\arguments{
\item{vcf_path}{Path to the input VCF file.}

\item{output_vcf}{Path for the output VCF file with imputed data.}

\item{batch_size}{Number of variants to process per batch (default: 1000).}

\item{maxiter}{Number of improvement iterations for the random forest algorithm (default: 10).}

\item{ntree}{Number of decision trees in the random forest (default: 100).}

\item{threads}{Number of threads used for computation (default: 1, or one less than available on the system).}

\item{write_log}{If TRUE, writes a log file of the process (advised for large datasets).}

\item{logfile}{Path to the log file, used if \code{write_log} is TRUE.}
}
\value{
Path to the output VCF file with imputed data.
}
\description{
Performs imputation of missing genomic data using a random forest algorithm, optimized for large-scale VCF files. This function reads VCF files in batches, applies random forest imputation via the missForest package, and writes the results to a new VCF file. The choice of batch size, number of improvement iterations, and the number of decision trees in the random forest is critical for balancing accuracy and computational demand.
}
\examples{
rfImputation("/path/to/input.vcf",
             "/path/to/output.vcf",
             batch_size = 1000,
             maxiter = 10,
             ntree = 100,
             threads = 5,
             write_log = TRUE,
             logfile = "/path/to/logfile.txt")

}
