% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gt_data.R
\name{seperateByPopulations}
\alias{seperateByPopulations}
\title{seperateByPopulations}
\usage{
seperateByPopulations(object, pop_assignments)
}
\arguments{
\item{object}{A S4 object of class vcfR.}

\item{pop_assignments}{A named vector. Elements are the population names and names are the individual name.}
}
\value{
A list containing one vcfR object per population.
}
\description{
Seperates a vcfR object into new objects per population. Needs to be done prior to calculating allele frequencies.
}
\examples{
mys <- c("8449", "8128", "8779", "8816", "8823", "8157")
dav <- c("8213", "8241", "8232", "8224", "10165", "8221", "8813", "8825", "8182", "8187")

individuals <- c(mys, dav)
pop_names <- c(rep("mys", length(mys)), rep("dav", length(dav)))
pop_assignments <- setNames(pop_names, individuals)

data("real", package = "GenoPop")
vcfs <- seperateByPopulations(real, pop_assignments)

}
