% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{TajimasD}
\alias{TajimasD}
\title{Tajima's D}
\usage{
TajimasD(object, seq_length)
}
\arguments{
\item{object}{A S4 object of type myVcfR. Allele frequencies and genotype matrix must be present.}

\item{seq_length}{Length of the sequence in number of bases. Must be provided to accurately work with all monomorphic sites, including those monomorphic for the reference, which are generally not included in a vcf.}
}
\value{
Tajima's D value.
}
\description{
Calculate Tajima's D statistic for a given dataset, a measure for neutrality.
}
\examples{
data("mys", package = "GenoPop")
TajimasD(mys, 265392)

}
