% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{TajimasD}
\alias{TajimasD}
\title{Tajima's D}
\usage{
TajimasD(object)
}
\arguments{
\item{object}{An S4 object of type myVcfR. Allele frequencies and genotype matrix must be present.}
}
\value{
Tajima's D value.
}
\description{
Calculate Tajima's D statistic for a given dataset, a measure for neutrality.
}
\examples{
data("mys", package = "GenoPop")
TajimasD(mys)

}
