% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{TwoDimSFS}
\alias{TwoDimSFS}
\title{TwoDimSFS}
\usage{
TwoDimSFS(objects, folded = FALSE)
}
\arguments{
\item{objects}{A list of two S4 objects of type myVcfR. Allele frequencies and genotype matrices must be present.}

\item{folded}{Logical, deciding if folded (TRUE) or unfolded (FALSE) SFS is returned. (Default is unfolded (FALSE).)}
}
\value{
Two-dimensional site frequency spectrum as a matrix
}
\description{
Calculate a two-dimensional site frequency spectrum from a list of two myVcfR objects.
}
\examples{
data("mys", package = "GenoPop")
data("dav", package = "GenoPop")
TwoDimSFS(list(mys, dav), folded = TRUE)

}
