% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_genomics.R
\name{WattersonsTheta}
\alias{WattersonsTheta}
\title{WattersonsTheta}
\usage{
WattersonsTheta(object, seq_length)
}
\arguments{
\item{object}{A S4 object of type GPvcfR. Allele frequencies and genotype matrix must be present.}

\item{seq_length}{The length of the sequence in the data set.}
}
\value{
Watterson's theta value.
}
\description{
Calculate Watterson's thea, a measure for neutrality, from an GPvcfR object. The metric will be normalized by the sequence length to make it comparable between data sets.
}
\examples{
data("mys", package = "GenoPop")
WattersonsTheta(mys, 265392)

}
